backend/django_app/requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
PyMySQL==1.1.2
sqlparse==0.5.3
cryptography>=41.0.0
pgeocode==0.4.1